Fixed debugging issue in arm image

+++ pwn-chal-arm/docker-entrypoint.sh
@@ -75,14 +75,6 @@ chattr +i "$FLAG_FILE" "/app/$CHAL_NAME" &>/dev/null
 ###### QEMU SETUP #######
 LIBRARY_PATH="/usr/arm-linux-gnueabihf";
 EMULATOR="qemu-arm"
-DEBUG=
-if [ ! -z "$QEMU_GDB_DEBUG" ]; then
-    if [ -z "$QEMU_GDB_PORT" ]; then
-        echo "[DEBUG] No QEMU_GDB_PORT specified. Defaulting to 1024"
-        QEMU_GDB_PORT=1024
-    fi
-    DEBUG=(-g $QEMU_GDB_PORT)
-fi
 
 ln -s "$LIBRARY_PATH/lib/ld-linux-armhf.so.3" "/usr/lib/ld-linux-armhf.so.3"
 ln -s "$LIBRARY_PATH/lib/ld-linux.so.3" "/lib/ld-linux.so.3" &>/dev/null
@@ -92,13 +84,25 @@ export LD_LIBRARY_PATH="$LIBRARY_PATH"
 echo "[QEMU] using $EMULATOR and libaries @ $LIBRARY_PATH"
 
 cd "$START_DIR";
+
+# Run the gdb server, host the binary.
+## NOTE: The binary will be hosted, but the stdin/out/err won't be redirected, so you'll have to communicate using docker :(
+# if someone can help me with this, that'll be helpful, ty.
+if [ ! -z "$QEMU_GDB_PORT" ]; then
+    echo "[GDB] Enabling QEMU's GDB remote debugging on $QEMU_GDB_PORT"
+    while [[ 1 ]]; do
+        LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -g $QEMU_GDB_PORT -L "$LIBRARY_PATH" "/app/$CHAL_NAME"
+    done
+    exit 0
+fi
+
 echo "Running $CHAL_NAME in $(pwd) as $RUN_AS using $BASE and listening locally on $PORT"
 if [ "$BASE" == "socat" ]; then
     rm -f /opt/ynetd
-    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR $DEBUG -L "$LIBRARY_PATH" su $RUN_AS -c "/opt/socat tcp-l:$PORT,reuseaddr,fork, EXEC:\"/app/$CHAL_NAME\",stderr | tee -a $LOG_FILE"
+    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -L "$LIBRARY_PATH" su $RUN_AS -c "/opt/socat tcp-l:$PORT,reuseaddr,fork, EXEC:\"/app/$CHAL_NAME\",stderr | tee -a $LOG_FILE"
 else
     rm -f /opt/socat
     # -lt => cpu time in seconds. Keeps connection opened for max 10 seconds.
     # -se => stderr to redirect to socket
-    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR $DEBUG -L "$LIBRARY_PATH" /opt/ynetd -lt 1 -p $PORT -u $RUN_AS -se y -d $START_DIR "/app/$CHAL_NAME" | tee -a $LOG_FILE
+    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -L "$LIBRARY_PATH" /opt/ynetd -lt 1 -p $PORT -u $RUN_AS -se y -d $START_DIR "/app/$CHAL_NAME" | tee -a $LOG_FILE
 fi
